requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
tqdm>=4.65.0
click>=8.0.0
open-clip-torch>=2.20.0
lpips>=0.1.4
